move checkbox to atoms
