+++ src/business/index.tsx
@@ -1,7 +1,6 @@
 export { Toggle } from './Toggle';
 export { MainMenu, MainMenuItem } from './MainMenu';
 export { Table, TableBodyCell, TableBodyRow, TableHeadCell, TablePagingControls, useTableState } from './Table';
-export { Checkbox } from './Checkbox';
 export { Tabs, Tab } from './Tabs';
 export { Button } from './Button';
 export { Tooltip } from './Tooltip';
