+++ src/index.ts
@@ -4,6 +4,7 @@ export { default as Box } from './atoms/Box';
 export { StatefulBox } from './atoms/Box';
 export { default as Button } from './atoms/Button';
 export { default as Caption } from './atoms/Caption';
+export { Checkbox } from './atoms/Checkbox';
 export { default as ColorPicker } from './atoms/ColorPicker';
 export { default as Container } from './atoms/Container';
 export { default as DescriptionList } from './atoms/DescriptionList';
